refactor: extract main-process icon and window services
Move icon extraction and running-window process capture utilities out of main.js into focused service modules while preserving existing IPC behavior and runtime output shapes.

Made-with: Cursor
